Apps: Add 02-loader project
This is a DFU loader which allows copying apps into the RAM and then
executes the app.

This will allow fast development of the main firmware without using up
the flash program cycles.

This project is a modified copy of
https://github.com/Solartraveler/UniversalboxArm/tree/a2017e5d17217898de555d77526e7ae967745359/src/apps/03-loader

It has been adjusted to the STM32F405,
the display option, key input and the option to load and store programs in an external memory have been removed.